Make minor improvements
Expand dataframe filter at the end of Stage 2 to filter out pages
with titles like "p. 2", "front", "", etc. in addition to those
that were already being filtered out.

In Step 4 changed subprocess.call to subprocess.run and set it up
to capture and print subprocess output. In the future, this
section should be changed to import ezid.py to avoid the unneeded
subprocess call.